+++ test/sandbox/app/components/cacheable_child_component.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class CacheableChildComponent < ViewComponent::Base
+  include ViewComponent::ExperimentallyCacheable
+end